# openapi.yaml
openapi: 3.1.0
info:
  title: EFB Connector API
  description: |
    Multi-tenant portal for syncing GPS tracks from Garmin Connect to the
    Kanu-EFB portal. Provides magic-link authentication, encrypted credential
    storage, and scheduled sync.
  version: 1.0.0
  contact:
    url: https://efb-connector.sauroter.de
servers:
  - url: https://efb-connector.sauroter.de
    description: Production
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Session cookie set after magic-link verification.
    internalBearer:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <INTERNAL_SECRET>` — constant-time validated.

  # -------------------------------------------------------------------------
  # Reusable schemas
  # -------------------------------------------------------------------------
...
paths:
...
  /login:
    get:
      tags: [Public]
      summary: Login form
      operationId: getLoginForm
      responses:
        "200":
          description: HTML login form
          content:
            text/html:
              schema:
                type: string
    post:
      tags: [Public]
      summary: Submit login (send magic link)
      description: |
        Sends a magic-link email to the provided address.
        Rate-limited by email + IP.
      operationId: postLogin
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required: [email]
      responses:
        "303":
          description: Redirect to confirmation page
        "400":
          description: Invalid email
        "429":
          description: Rate limit exceeded
...
  /dashboard:
    get:
      tags: [Dashboard]
      summary: Dashboard
      description: |
        Shows connection status, last sync stats, and a setup wizard
        if the user has not completed initial configuration.
      operationId: getDashboard
      security:
        - sessionCookie: []
      responses:
        "200":
          description: HTML dashboard
          content:
            text/html:
              schema:
                type: string
        "303":
          description: Redirect to /login if unauthenticated
  /settings:
    get:
      tags: [Dashboard]
      summary: Settings page
      description: Consolidated settings view with tabs for Garmin, EFB, and preferences.
      operationId: getSettings
      security:
        - sessionCookie: []
      responses:
        "200":
          description: HTML settings page
          content:
            text/html:
              schema:
                type: string
  /settings/garmin:
    get:
      tags: [Dashboard]
      summary: Garmin settings form
      operationId: getGarminSettings
      security:
        - sessionCookie: []
      responses:
        "200":
          description: HTML form for Garmin credentials
          content:
            text/html:
              schema:
                type: string
    post:
      tags: [Dashboard]
      summary: Save Garmin credentials
      description: |
        Validates credentials against Garmin Connect, then encrypts and stores them.
        Handles MFA errors.
      operationId: postGarminSettings
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                csrf_token:
                  type: string
              required: [email, password, csrf_token]
      responses:
        "303":
          description: Redirect to /dashboard on success or /settings/garmin on error
        "400":
          description: Missing required fields
  /settings/garmin/mfa:
    get:
      tags: [Dashboard]
      summary: Garmin MFA code entry form
      description: |
        Renders a form for entering the 6-digit MFA verification code.
        Redirects to /settings/garmin if no MFA session is pending.
      operationId: getGarminMFA
      security:
        - sessionCookie: []
      responses:
        "200":
          description: HTML form for MFA code entry
          content:
            text/html:
              schema:
                type: string
        "303":
          description: Redirect to /settings/garmin if no MFA session
    post:
      tags: [Dashboard]
      summary: Submit Garmin MFA code
      description: |
        Completes the MFA verification started during credential validation.
        On success, marks credentials as valid and redirects to /dashboard.
        On failure, redirects to /settings/garmin with an error flash.
      operationId: postGarminMFA
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                mfa_code:
                  type: string
                  pattern: "^[0-9]{6}$"
                  description: 6-digit verification code from authenticator app
                csrf_token:
                  type: string
              required: [mfa_code, csrf_token]
      responses:
        "303":
          description: Redirect to /dashboard on success or /settings/garmin on error
        "400":
          description: Missing required fields
...
  /settings/efb:
    get:
      tags: [Dashboard]
      summary: EFB settings form
      operationId: getEFBSettings
      security:
        - sessionCookie: []
      responses:
        "200":
          description: HTML form for EFB credentials
          content:
            text/html:
              schema:
                type: string
    post:
      tags: [Dashboard]
      summary: Save EFB credentials
      description: Validates credentials against the EFB portal, then encrypts and stores them.
      operationId: postEFBSettings
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                  format: password
                csrf_token:
                  type: string
              required: [username, password, csrf_token]
      responses:
        "303":
          description: Redirect to /dashboard on success or /settings/efb on error
        "400":
          description: Missing required fields
...
  /sync/history:
    get:
      tags: [Dashboard]
      summary: Sync history
      operationId: getSyncHistory
      security:
        - sessionCookie: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: HTML page listing sync runs
          content:
            text/html:
              schema:
                type: string
...